[backend] Move db updater stuff into own package
