+++ backend/backend/src/main/java/de/aaaaaaah/velcom/backend/listener/dbupdate/DbUpdateException.java
@@ -1,4 +1,4 @@
-package de.aaaaaaah.velcom.backend.listener;
+package de.aaaaaaah.velcom.backend.listener.dbupdate;
 
 public class DbUpdateException extends Exception {
